// server/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Provider = {
  GOOGLE: 'GOOGLE'
} as const

export type Provider = (typeof Provider)[keyof typeof Provider]


export const Asset = {
  SOL: 'SOL',
  USDT: 'USDT'
} as const

export type Asset = (typeof Asset)[keyof typeof Asset]


export const OrderSide = {
  BUY: 'BUY',
  SELL: 'SELL'
} as const

export type OrderSide = (typeof OrderSide)[keyof typeof OrderSide]


export const OrderType = {
  LIMIT: 'LIMIT',
  MARKET: 'MARKET'
} as const

export type OrderType = (typeof OrderType)[keyof typeof OrderType]


export const OrderStatus = {
  OPEN: 'OPEN',
  PARTIALLY_FILLED: 'PARTIALLY_FILLED',
  FILLED: 'FILLED',
  CANCELLED: 'CANCELLED',
  REJECTED: 'REJECTED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const CreatedType = {
  HUMAN: 'HUMAN',
  SYSTEM: 'SYSTEM'
} as const

export type CreatedType = (typeof CreatedType)[keyof typeof CreatedType]


export const LedgerType = {
  DEPOSIT: 'DEPOSIT',
  TRADE: 'TRADE',
  FEE: 'FEE',
  RESERVATION: 'RESERVATION',
  RELEASE: 'RELEASE'
} as const

export type LedgerType = (typeof LedgerType)[keyof typeof LedgerType]